// Frontend/jump-park-rh/src/types/apps/recapTypes.ts
export type RecapType = {
  id: number
  matricule: string
  full_name: string
  nbHeureTrav: string
  nbJourTrav: string
  nbJourSup: string
  HeureSupplimentaire: string
  conge: string
  absance: string
  autorisation: string
  jourFerie: string
  service: string
  department_id: string
  department: string
}